dojo16-questionario/@ingridDuarte, @nicoledbianchin, @eduardaschuler, @gabrierabelo: separa perguntas da criacao do diagnostico
